requirements.txt
alembic==1.4.2
fastapi[all]==0.47.1
flake8==3.8.3
pre-commit==1.18.2
psycopg2-binary==2.8.4
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
jupyterlab==0.35.4
Keras==2.2.4
matplotlib==3.0.0
networkx==2.2
numpy==1.15.2
pandas==0.23.4
python==3.6.6
python-utils==2.3.0
scipy==1.1.0
seaborn==0.9.0
SQLAlchemy==1.3.3
tweepy==3.9.0